Allow to use the Wi-Fi mac address to find network location aliases
The current solution to get the current SSID relies on the fact that
macOS rearranges the preferred networks depending on when they were
last used. This appears to work most of the time, but there is a non
zero chance Apple will change this to further protect the privacy
of the users.

A more reliable way to detect the current network is that of reading
the MAC address used to connect to some Wi-Fi network. Each saved
network get by default a unique, random and fixed MAC address. This
allows to map Wi-Fi networks to network locations.

Given that naming network locations using a MAC address looks awkward,
the mapping with the MAC address can only be done via alias.conf.