Allow to use the Wi-Fi mac address to find network location aliases
Due to privacy reasons macOS hides the SSID of the current network.
root can still see the network, while regular users can only guess
it using workarounds such as looking at the order of the preferred
network list, which is not reliable and might not work in future.

A more reliable way to detect the current network is that of reading
the MAC address used to connect to some Wi-Fi network. Each saved
network gets by default a unique, random and fixed MAC address. This
allows to map Wi-Fi networks to network locations.

Given that naming network locations using a MAC address looks awkward,
the mapping with the MAC address can only be done via alias.conf.
SSID aliases take precedence.

Note: the alias parser now discards all the lines not starting with
a SSID immediately followed by `=`. This allows to write comments
in the file, useful when writing aliases for MAC addresses.